EditDialog: refactor - move context
