+++ src/components/FeaturePanel/EditDialog/EditContent/FeatureEditSection/ItemHeading.tsx
@@ -2,7 +2,7 @@ import { Stack, Typography } from '@mui/material';
 import { getOsmTypeFromShortId, NwrIcon } from '../../../NwrIcon';
 import React from 'react';
 import styled from '@emotion/styled';
-import { useCurrentItem } from '../../EditContext';
+import { useCurrentItem } from '../../context/EditContext';
 
 const StyledTypography = styled(Typography, {
   shouldForwardProp: (prop) => !prop.startsWith('$'),
