+++ src/services/osm/auth/__tests__/getDiffXml.test.ts
@@ -1,4 +1,4 @@
-import { DataItem } from '../../../../components/FeaturePanel/EditDialog/useEditItems';
+import { DataItem } from '../../../../components/FeaturePanel/EditDialog/context/useEditItems';
 import { getDiffXml } from '../getDIffXml';
 
 const nodeNew = {
