+++ src/services/osm/auth/getDIffXml.ts
@@ -1,4 +1,4 @@
-import { DataItem } from '../../../components/FeaturePanel/EditDialog/useEditItems';
+import { DataItem } from '../../../components/FeaturePanel/EditDialog/context/useEditItems';
 import { getApiId } from '../../helpers';
 import {
   DiffDocXmljs,
